face_recog/python_backend/requirements.txt
setuptools>=65.5.1
wheel>=0.38.4
torch==2.1.2
torchvision==0.16.2
ultralytics==8.2.0
opencv-python==4.9.0.80
numpy==1.24.4
Pillow==10.2.0
flask==2.3.3
flask-cors==4.0.0
mtcnn==0.1.1
scikit-learn==1.3.2
scipy==1.11.4
tensorflow==2.15.0
keras==2.15.0
matplotlib==3.8.3
albumentations==1.3.1  # Image augmentation
timm==0.9.12  # Additional model architectures
